+++ src/js/profiling/hermes.ts
@@ -1,3 +1,7 @@
+import { Platform } from 'react-native';
+
+import { ANDROID_DEFAULT_BUNDLE_NAME, IOS_DEFAULT_BUNDLE_NAME } from '../integrations/rewriteframes';
+
 export type StackFrameId = number;
 export type MicrosecondsSinceBoot = string;
 
@@ -24,10 +28,17 @@ export interface Sample {
 }
 
 export interface StackFrame {
+  // Hermes Bytecode
+  funcVirtAddr?: string;
+  offset?: string;
+
+  // JavaScript
   line?: string;
   column?: string;
   funcLine?: string;
   funcColumn?: string;
+
+  // Common
   name: string;
   category: string;
   parent?: number;
@@ -39,13 +50,5 @@ export interface Profile {
   stackFrames: Record<string, StackFrame>;
 }
 
-/**
- * Hermes Profile Stack Frame Name contains function name and file path.
- *
- * `foo(/path/to/file.js:1:2)` -> `foo`
- */
-export function parseHermesStackFrameFunctionName(hermesName: string): string {
-  const indexOfLeftParenthesis = hermesName.indexOf('(');
-  const name = indexOfLeftParenthesis !== -1 ? hermesName.substring(0, indexOfLeftParenthesis) : hermesName;
-  return name;
-}
+export const DEFAULT_BUNDLE_NAME =
+  Platform.OS === 'android' ? ANDROID_DEFAULT_BUNDLE_NAME : Platform.OS === 'ios' ? IOS_DEFAULT_BUNDLE_NAME : undefined;