+++ src/js/profiling/types.ts
@@ -1,8 +1,13 @@
-import type { Profile, ThreadCpuProfile } from '@sentry/types';
+import type { Profile, ThreadCpuFrame as SentryThreadCpuFrame, ThreadCpuProfile } from '@sentry/types';
 
 import type { NativeProfileEvent } from './nativeTypes';
 
+export interface ThreadCpuFrame extends SentryThreadCpuFrame {
+  in_app?: boolean;
+}
+
 export interface RawThreadCpuProfile extends ThreadCpuProfile {
+  frames: ThreadCpuFrame[];
   profile_id?: string;
   active_thread_id: string;
 }